backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Not needed for SQLite
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.3.7
httpx==0.24.1
python-jobspy==1.1.82
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
reportlab==4.0.4
email-validator==2.1.0
psutil==5.9.6
python-magic==0.4.27
